DED_AI_Model_Server/requirements.txt
flask
gunicorn
torch
segmentation-models-pytorch
numpy
opencv-python
Pillow
six